pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask-login>=0.6.3",
    "flask-wtf>=1.2.2",
    "google-generativeai>=0.8.4",
    "oauthlib>=3.2.2",
    "twilio>=9.4.4",
]
